# warehouse-svc/build-warehouse/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BookID": {
				"type": "string"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "TSOA",
		"version": "1.0.0",
		"description": "Build swagger-compliant REST APIs using TypeScript and Node",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "unknown"
		}
	},
	"paths": {
		"/warehouse/{bookId}": {
			"get": {
				"operationId": "GetBookOnShelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/BookID"
						}
					}
				]
			}
		},
		"/warehouse": {
			"post": {
				"operationId": "CreateOrUpdateBookStock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"numberOfBooks": {
										"type": "number",
										"format": "double"
									},
									"bookId": {
										"$ref": "#/components/schemas/BookID"
									},
									"shelf": {
										"type": "string"
									}
								},
								"required": [
									"numberOfBooks",
									"bookId",
									"shelf"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/warehouse/fulfill/{orderId}": {
			"put": {
				"operationId": "FulfillOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "number",
										"format": "double"
									},
									"bookId": {
										"$ref": "#/components/schemas/BookID"
									}
								},
								"required": [
									"quantity",
									"bookId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}